Revert "refactor: Move BB console commands out of Bladeburner.ts"
This reverts commit 640e9cf7102c36d1004ef9ee3e9c1e7c6a833c75.
